refactor: validate로직 추출
