+++ src/utils/card.ts
@@ -22,3 +22,27 @@ export const getFormattedValidityPeriodUnit = (
   const { month, year } = validityPeriod;
   return `${month ? month + "/" : ""}${year ? year : ""}`;
 };
+
+export const formatValidityPeriod = (nextRaw: string, prev: string) => {
+  const onlyNumber = nextRaw.replace(/\D/g, "");
+
+  if (onlyNumber === "") return "";
+
+  if (onlyNumber.length === 1) {
+    if (onlyNumber === "0") return "0";
+
+    return `0${onlyNumber}`;
+  }
+
+  if (prev === "0") {
+    if (onlyNumber === "00") return "0";
+
+    return onlyNumber.slice(0, 2);
+  }
+
+  if (onlyNumber.startsWith("0")) {
+    return onlyNumber.slice(1, 3);
+  }
+
+  return onlyNumber.slice(0, 2);
+};
